Change byte to int

+++ Entities/Card.cs
@@ -4,9 +4,8 @@ namespace DurakGame.Entities;
 
 public class Card
 {
-    public byte Id { get; set; }
+    public int Id { get; set; }
     public Rank Rank { get; set; }
-
     public Suit Suit { get; set; }
     
 }
